+++ machines/nfa/nfa.go
@@ -8,7 +8,7 @@ import (
 	"strconv"
 )
 
-var count int = 0
+var count int64 = 0
 
 type Machine struct {
 	Start *State
